pharmaverseadam: ADaM Test Data for the 'Pharmaverse' Family of Packages

A set of Analysis Data Model (ADaM) datasets constructed using the Study Data Tabulation Model (SDTM) datasets contained in the 'pharmaversesdtm' package and the template scripts from the 'admiral' family of packages. ADaM dataset specifications are described in the CDISC ADaM implementation guide, accessible by creating a free account on https://www.cdisc.org/.

See also

Author

Maintainer: Fanny Gautier fanny.gautier@cytel.com (ORCID)

Authors:

Stefan Bundfuss (ORCID)

Edoardo Mancini edoardo.mancini@roche.com (ORCID)

Lina Patil lina.patil@cytel.com

Vladyslav Shuliar (ORCID)

Other contributors:

Cytel Inc. [copyright holder, funder]

F. Hoffmann-La Roche AG [copyright holder, funder]

GlaxoSmithKline LLC [copyright holder, funder]